extra download now happens with setup
