+++ .github/workflows/test.yml
@@ -30,7 +30,9 @@ jobs:
       run: |
         pip install pylint
         pip install -e .
-        python download_extras.py
+        export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/path/to/mujoco/lib
+        export MJLIB_PATH=/path/to/home/.mujoco/mujoco210/bin/libmujoco210.so
+        export MUJOCO_PY_MUJOCO_PATH=/path/to/home/.mujoco/mujoco210
 
     - name: Check code quality
       run: |
